use proper type

+++ libsel4allocman/src/bootstrap.c
@@ -1111,8 +1111,8 @@ int allocman_add_simple_untypeds_with_regions(allocman_t *alloc, simple_t *simpl
     int error = prepare_handle_device_untyped_cap(alloc, simple, &state, num_regions, region_list);
     ZF_LOGF_IF(error, "bootstrap_prepare_handle_device_untyped_cap Failed");
 
-    size_t i;
-    size_t total_untyped = simple_get_untyped_count(simple);
+    int i;
+    int total_untyped = simple_get_untyped_count(simple);
 
     for(i = 0; i < total_untyped; i++) {
         size_t size_bits;
